% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graticule-package.R
\docType{package}
\name{graticule-package}
\alias{graticule-package}
\alias{_PACKAGE}
\title{graticule: Meridional and Parallel Lines for Maps}
\description{
Create graticule lines and labels for maps. Control the creation of lines or tiles by setting their placement (at particular meridians and parallels) and extent (along parallels and meridians). Labels are created independently of lines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hypertidy/graticule}
  \item Report bugs at \url{https://github.com/hypertidy/graticule/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
