% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graticule.R
\name{graticule}
\alias{graticule}
\title{Create a graticule.}
\usage{
graticule(easts, norths, ndiscr = 60, xlim, ylim)
}
\arguments{
\item{easts}{longitudes for meridional lines}

\item{norths}{latitudes for parallel lines}

\item{ndiscr}{number of discrete vertices for each segment}

\item{xlim}{maximum range of parallel lines}

\item{ylim}{maximum range of meridional lines}
}
\description{
Create a graticule.
}
\examples{
\dontrun{
library(rgdal)
x <- as.matrix(expand.grid(x = seq(100, 240, by = 15), y = seq(-85, -30, by = 15)))
prj <- "+proj=laea +lon_0=180 +lat_0=-70"
px <- project(x, prj)
g <- graticule(unique(x[,1]), unique(x[,2]))
pg <- spTransform(g, CRS(prj))
plot(px, type = "n")
plot(pg, add = TRUE)

g2 <- graticule(unique(x[,1]), unique(x[,2]), ylim = c(-90, 0), xlim = c(110, 250))
pg2 <- spTransform(g2, CRS(prj))
plot(px, type = "n")
plot(pg2, add = TRUE)

}
}

