% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graticule.R
\name{graticule_labels}
\alias{graticule_labels}
\title{Create graticule labels.}
\usage{
graticule_labels(easts, norths, xlines, ylines, proj = NULL)
}
\arguments{
\item{easts}{longitudes for meridional labels}

\item{norths}{latitudes for parallel labels}

\item{xlines}{meridian/s for placement of parallel labels}

\item{ylines}{parallel/s for placement of meridian labels}
}
\description{
Create graticule labels.
}
\examples{
plot(graticule(easts = c(100, 120, 160, 180), norths = c(-80,-70,-60, -50,-45, -30),  proj = "+proj=lcc +lon_0=150 +lat_0=-80"))
op <- par(xpd = NA); text(labs, lab = parse(text = labs$lab), pos = c(2, 1)[labs$islon + 1], adj = 1.2);par(op)
}

