% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graticule2.R
\name{graticule_tiles}
\alias{graticule_tiles}
\title{@importFrom stats approx
 @importFrom utils head}
\usage{
graticule_tiles(lons = seq(-180, 180, by = 15), lats = seq(-84, 84, by
  = 12), nverts = 24, proj = NULL, margin = FALSE)
}
\description{
@importFrom stats approx
 @importFrom utils head
}
