% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.r
\name{rgrat}
\alias{rgrat}
\title{Generate random graticules.}
\usage{
rgrat(family = NULL, lon_0 = NULL, lat_0 = NULL, radius = NULL,
  datum = NULL, ...)
}
\arguments{
\item{family}{central longitude}

\item{...}{}
}
\value{
see \code{\link{auto_graticule}}
}
\description{
All arguments are optional and will be auto sampled sensibly if not specified.
}

